refactor: Move fixtures and enum to proper dest
There is also preparation for removal of Managers
