+++ app/src/main/java/com/dreamtracker/app/habit/domain/model/HabitFrequency.java
@@ -1,4 +1,4 @@
-package com.dreamtracker.app.habit.domain.utils;
+package com.dreamtracker.app.habit.domain.model;
 
 public enum HabitFrequency {
     DAILY,WEEKLY,MONTHLY
