+++ app/src/main/java/com/dreamtracker/app/habit/domain/model/HabitStatus.java
@@ -1,4 +1,4 @@
-package com.dreamtracker.app.habit.domain.utils;
+package com.dreamtracker.app.habit.domain.model;
 
 public enum HabitStatus {
     ACTIVE,INACTIVE
