+++ app/src/test/java/com/dreamtracker/app/fixtures/UserFixtures.java
@@ -1,4 +1,4 @@
-package com.dreamtracker.app.habit.domain.fixtures;
+package com.dreamtracker.app.fixtures;
 
 import com.dreamtracker.app.user.domain.model.User;
