+++ app/src/test/java/com/dreamtracker/app/habit/adapters/api/CategoryControllerTest.java
@@ -1,10 +1,9 @@
 package com.dreamtracker.app.habit.adapters.api;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import com.dreamtracker.app.configuration.TestPostgresConfiguration;
-import com.dreamtracker.app.habit.domain.fixtures.CategoryFixtures;
+import com.dreamtracker.app.fixtures.CategoryFixtures;
 import com.dreamtracker.app.infrastructure.response.Page;
 import com.dreamtracker.app.user.config.CurrentUserProvider;
 import com.dreamtracker.app.user.config.MockCurrentUserProviderImpl;
