+++ app/src/test/java/com/dreamtracker/app/infrastructure/repository/SpringDataHabitRepositoryTest.java
@@ -1,10 +1,9 @@
 package com.dreamtracker.app.infrastructure.repository;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import com.dreamtracker.app.configuration.TestPostgresConfiguration;
-import com.dreamtracker.app.habit.domain.fixtures.HabitFixture;
+import com.dreamtracker.app.fixtures.HabitFixture;
 import com.dreamtracker.app.user.config.CurrentUserProvider;
 import com.dreamtracker.app.user.config.MockCurrentUserProviderImpl;
 import org.junit.jupiter.api.BeforeEach;
