+++ app/src/test/java/com/dreamtracker/app/infrastructure/repository/SpringDataHabitTrackRepositoryTest.java
@@ -3,8 +3,8 @@ package com.dreamtracker.app.infrastructure.repository;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import com.dreamtracker.app.configuration.TestPostgresConfiguration;
-import com.dreamtracker.app.habit.domain.fixtures.HabitFixture;
-import com.dreamtracker.app.habit.domain.fixtures.HabitTrackFixture;
+import com.dreamtracker.app.fixtures.HabitFixture;
+import com.dreamtracker.app.fixtures.HabitTrackFixture;
 import com.dreamtracker.app.habit.domain.model.Habit;
 import com.dreamtracker.app.user.config.CurrentUserProvider;
 import com.dreamtracker.app.user.config.MockCurrentUserProviderImpl;
